remove time_context from util
